Moved the apio/proto directory to apio/common since it's used by both the apio process and the scons child process.
