+++ apio/commands/apio_test.py
@@ -13,7 +13,7 @@ import click
 from apio.managers.scons import SCons
 from apio.commands import options
 from apio.apio_context import ApioContext, ApioContextScope
-from apio.proto.apio_pb2 import ApioTestParams
+from apio.common.proto.apio_pb2 import ApioTestParams
 
 
 # ---------------------------
