+++ test/scons/testing.py
@@ -9,7 +9,7 @@ import SCons.Defaults
 import SCons.Script.Main
 from google.protobuf import text_format
 from apio.scons.apio_env import ApioEnv
-from apio.proto.apio_pb2 import SconsParams, TargetParams
+from apio.common.proto.apio_pb2 import SconsParams, TargetParams
 
 # R0801: Similar lines in 2 files
 # pylint: disable=R0801
