+++ tox.ini
@@ -68,8 +68,8 @@ setenv=
 # to use the definitions in the .pyi stubs istead of the criptic protocol
 # buffers .py files.
 commands = 
-    black   {env:LINT_ITEMS}  --exclude apio/proto
-    flake8  {env:LINT_ITEMS}  --exclude apio/proto
+    black   {env:LINT_ITEMS}  --exclude apio/common/proto
+    flake8  {env:LINT_ITEMS}  --exclude apio/common/proto
     pylint  {env:LINT_ITEMS}  --prefer-stubs True 
 
 # ----------------------------------------------------
